feat: 위치에 대한 책임을 분리한 BoardPosition 클래스 구현
- 위치 문자열 파라미터를 받는 생성자를 통해 List의 position을 계산하는 기능 구현
